Admin could see latest access log

+++ include/pageboot.php
@@ -65,39 +65,42 @@ $db->cache = Zend\Cache\StorageFactory::factory($config->cache->toArray());
    }
 }
 $db->config = $config;
-$logName = __DIR__.'/../log';
-if(!is_dir($logName)) {
-   mkdir($logName);
-}
-$logName .= '/'.date('Y-m').'.csv';
-$s = array();
 
-$s['HTTP_USER_AGENT']=$_SERVER['HTTP_USER_AGENT'];
-$s['REMOTE_ADDR']='';
-if (key_exists('HTTP_X_FORWARDED_FOR',$_SERVER)) {
-$s['REMOTE_ADDR']=$_SERVER['HTTP_X_FORWARDED_FOR'];
+if (is_array($_REQUEST) && !key_exists('no_log',$_REQUEST)) {
+    $logName = __DIR__.'/../log';
+    if(!is_dir($logName)) {
+       mkdir($logName);
+    }
+    $logName .= '/'.date('Y-m').'.csv';
+    $s = array();
+    
+    $s['HTTP_USER_AGENT']=$_SERVER['HTTP_USER_AGENT'];
+    $s['REMOTE_ADDR']='';
+    if (key_exists('HTTP_X_FORWARDED_FOR',$_SERVER)) {
+    $s['REMOTE_ADDR']=$_SERVER['HTTP_X_FORWARDED_FOR'];
+    }
+    if ($s['REMOTE_ADDR']=='') {
+       $s['REMOTE_ADDR']=$_SERVER['REMOTE_ADDR'];
+    }
+    $s['REQUEST_METHOD']=$_SERVER['REQUEST_METHOD'];
+    $s['REQUEST_URI']=$_SERVER['REQUEST_URI'];
+    $s['REQUEST_TIME']=date('Y-m-d H:i:s',$_SERVER['REQUEST_TIME']);
+    $s['HTTP_REFERER']='';
+    if (key_exists('HTTP_REFERER',$_SERVER)) {
+        $s['HTTP_REFERER']=$_SERVER['HTTP_REFERER'];
+    }
+    $s['SESSIONID']='';
+    if (isset($_COOKIE) && is_array($_COOKIE) && key_exists('abbrevia',$_COOKIE)) {
+       $s['SESSIONID']=$_COOKIE['abbrevia'];
+    }
+    $s = array_map(function($val) {
+        return str_replace(',','',$val);
+    },$s);
+    if (!is_file($logName) || filesize($logName)==0) {
+       file_put_contents($logName,implode(',',array_keys($s)).PHP_EOL);
+    }
+    file_put_contents($logName,implode(',',$s).PHP_EOL,FILE_APPEND);
 }
-if ($s['REMOTE_ADDR']=='') {
-   $s['REMOTE_ADDR']=$_SERVER['REMOTE_ADDR'];
-}
-$s['REQUEST_METHOD']=$_SERVER['REQUEST_METHOD'];
-$s['REQUEST_URI']=$_SERVER['REQUEST_URI'];
-$s['REQUEST_TIME']=date('Y-m-d H:i:s',$_SERVER['REQUEST_TIME']);
-$s['HTTP_REFERER']='';
-if (key_exists('HTTP_REFERER',$_SERVER)) {
-    $s['HTTP_REFERER']=$_SERVER['HTTP_REFERER'];
-}
-$s['SESSIONID']='';
-if (isset($_COOKIE) && is_array($_COOKIE) && key_exists('abbrevia',$_COOKIE)) {
-   $s['SESSIONID']=$_COOKIE['abbrevia'];
-}
-$s = array_map(function($val) {
-return str_replace(',','',$val);
-},$s);
-if (!is_file($logName) || filesize($logName)==0) {
-   file_put_contents($logName,implode(',',array_keys($s)).PHP_EOL);
-}
-file_put_contents($logName,implode(',',$s).PHP_EOL,FILE_APPEND);
 require __DIR__.'/../lib/floraobservation/Autoload.php';
 floraobservation\Autoload::getInstance();
